requirements.txt
# --- Core runtime ---
python-dotenv>=1.0.0
requests>=2.31.0

# --- Agent / LLM ---
langchain>=1.2.0
langchain-community>=0.4.0
langchain-openai>=0.1.7
openai>=2.0.0

# --- Web (optional demo UI) ---
flask>=3.0.0

# --- Analytics (deterministic fallback + metrics) ---
numpy>=1.26.0

# --- Testing ---
pytest>=8.0.0
